refactor: switch to internal app package from main
This commit moves all the logic in the main package
to a new internal app package and adds a new simpler main.go
to the cmd/app main package.

With this refactor we remove the global state during the init and
scoping the global vars that we had earlier. This might help with
testing (if needed) later on as well.

The biggest benefit of this refactor for new users and developers
is running `go run cmd/app/main.go` to get started with listmonk.