Begin refactoring *Queries Stmts into methods
To put database calls behind an interface and eventually allow implementations
for multiple databases, I am slowly abstracting the raw Postgres calls in the
Queries struct into a form that can be implemented for another database driver.

In this commit, I moved several raw *sqlx.Stmt variables behind function calls
so that we can eventually make Queries an interface. To make the changes minimally
disruptive I maintained the same calling convention as the old code and attemtped
to change as little as possible.

A future commit will convert the remaining SQL statements that couldn't be converted
either because they were passed to helper functions in a complicated way or
because they are parametric.