The goal was to improve the processing of bounce emails.
In case the message bounces between the smtp servers, it should be processed in multipart format. This means that there can be multiple headers in the message and you have to dig down to the bottom to find the Campaign and Subscriber IDs and the other header variables.

In addition, I have found that the header is sometimes broken (when it contains a new line, for example) and can only be processed halfway. In this case, there is no other way than regular expressions to find all the required header IDs in the body (not just the Campaign and Subscriber IDs).
While reading the email body, I was surprised to find that after using "message.Read", the buffer pointer is no longer in the first position, so I need to reset the pointer first in order to get the full body (b, _ = c.RetrRaw ( id)).