docs: suggest an FQDN in docker compose file (#2019)
The 'hostname' field in the docker-compose file is later used in Listmonk to generate the 'Message-Id' of each mail. The hostname is used as the id-right section of 'Message-Id'.

According to RFC 2822 3.6.4, id-right = dot-atom-text / no-fold-literal / obs-id-right 
Also, Rspamd has a spam filtering field 'MID_RHS_NOT_FQDN' which defaults to a 1.0 score.

This commit suggests an FQDN in the docker-compose file, which is guaranteed to meet the requirement.